WIP: xwayland: Handle wl_tablet events
Adds support for the wl_tablet interface. Three input devices exist to
handle these events 'stylus', 'eraser', and 'cursor' (matching the
behavior of the Wacom driver). These devices are braindead and need to
be wired up to actually handle the events.

At the moment this is patch checks if a 'wl_tablet_manager' interface
is in the global object registry since that's the place you'll typically
run into it (e.g. under Weston or mutter). This isn't the ideal end-state
though -- ideally the seat should provide access to the tablet manager,
but that requires modifying the upstream Wayland interfaces and modifying
the compositors to follow suit. In the future when tablet support is
merged into Wayland itself, we'll use the (disabled) codepath to get the
manager from the current seat.